Improve the summary of an event, specially the dates/time zone
Was showing the dates using the user's time zone but was also printing the
time zone used to create the event. Was confusing.
Not it still shows the dates in the user's time zone, but prints right below
which is this time zone in which the dates are being shown. We don't show
anymore the time zone in which the event was created.

Also made the entire summary and the social links a bit nicer.

refs #1226